+++ src/Exception/FileNotExistsException.php
@@ -49,7 +49,7 @@ class FileNotExistsException extends Exception
     /**
      * Gets the path of file that do not exist
      * @return string|null
-     * @since 1.2.10
+     * @since 1.2.11
      * @uses $path
      */
     public function getFilePath(): ?string