+++ src/Exception/PropertyNotExistsException.php
@@ -22,13 +22,38 @@ use Exception;
 class PropertyNotExistsException extends Exception
 {
     /**
+     * @var string|null
+     */
+    protected $property;
+
+    /**
      * Constructor
-     * @param string $message The string of the error message
+     * @param string|null $message The string of the error message
      * @param int $code The code of the error
      * @param \Throwable|null $previous the previous exception
+     * @param string|null $property Name of the property that do not exist
+     * @uses $property
      */
-    public function __construct(string $message = 'Property does not exist', int $code = 0, ?\Throwable $previous = null)
+    public function __construct(?string $message = null, int $code = 0, ?\Throwable $previous = null, ?string $property = null)
     {
+        if (!$message) {
+            $message = 'Property does not exist';
+            if ($property) {
+                $message = sprintf('Property `%s` does not exist', $property);
+            }
+        }
         parent::__construct($message, $code, $previous);
+        $this->property = $property;
+    }
+
+    /**
+     * Gets the name of the property that do not exist
+     * @return string|null
+     * @since 1.2.11
+     * @uses $property
+     */
+    public function getPropertyName(): ?string
+    {
+        return $this->property;
     }
 }