+++ tests/TestCase/Exception/NotPositiveExceptionTest.php
@@ -1,0 +1,51 @@
+<?php
+declare(strict_types=1);
+/**
+ * This file is part of php-tools.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/php-tools
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ */
+namespace Tools\Test\Exception;
+
+use Tools\Exception\NotPositiveException;
+use Tools\TestSuite\TestCase;
+
+/**
+ * NotPositiveExceptionTest class
+ */
+class NotPositiveExceptionTest extends TestCase
+{
+    /**
+     * Test for the exception
+     * @test
+     */
+    public function testException()
+    {
+        try {
+            throw new NotPositiveException(null, 0, null, -1);
+        } catch (NotPositiveException $e) {
+            $this->assertSame('Value `-1` is not a positive', $e->getMessage());
+            $this->assertSame(-1, $e->getValue());
+        }
+
+        try {
+            throw new NotPositiveException();
+        } catch (NotPositiveException $e) {
+            $this->assertSame('Value is not a positive', $e->getMessage());
+            $this->assertNull($e->getValue());
+        }
+
+        try {
+            throw new NotPositiveException(null, 0, null, ['no-stringable']);
+        } catch (NotPositiveException $e) {
+            $this->assertSame('Value is not a positive', $e->getMessage());
+            $this->assertSame(['no-stringable'], $e->getValue());
+        }
+    }
+}